refactor: uniformize inventory header
